refactor feature home to quote
