+++ lib/core/dependency_injector.dart
@@ -1,12 +1,12 @@
 import 'package:challenge_app/core/data/http_client.dart';
 import 'package:challenge_app/core/providers/loading_provider.dart';
 import 'package:challenge_app/core/providers/theme_provider.dart';
-import 'package:challenge_app/features/home/data/repositories/quote_repository_impl.dart';
-import 'package:challenge_app/features/home/data/services/quote_service_impl.dart';
-import 'package:challenge_app/features/home/domain/repositories/quote_repository.dart';
-import 'package:challenge_app/features/home/domain/services/quote_service.dart';
-import 'package:challenge_app/features/home/domain/usecases/quote_usecase.dart';
-import 'package:challenge_app/features/home/presentation/providers/quote_provider.dart';
+import 'package:challenge_app/features/quote/data/repositories/quote_repository_impl.dart';
+import 'package:challenge_app/features/quote/data/services/quote_service_impl.dart';
+import 'package:challenge_app/features/quote/domain/repositories/quote_repository.dart';
+import 'package:challenge_app/features/quote/domain/services/quote_service.dart';
+import 'package:challenge_app/features/quote/domain/usecases/quote_usecase.dart';
+import 'package:challenge_app/features/quote/presentation/providers/quote_provider.dart';
 import 'package:challenge_app/features/login/data/repositories/login_repository_imp.dart';
 import 'package:challenge_app/features/login/data/services/login_service_imp.dart';
 import 'package:challenge_app/features/login/domain/repositories/login_repository.dart';
