+++ lib/core/routes/route_generator.dart
@@ -1,5 +1,5 @@
 import 'package:challenge_app/core/routes/app_routes.dart';
-import 'package:challenge_app/features/home/presentation/home_page.dart';
+import 'package:challenge_app/features/quote/presentation/quote_page.dart';
 import 'package:challenge_app/features/login/presentation/login_page.dart';
 import 'package:flutter/material.dart';
 
@@ -9,7 +9,7 @@ class RouteGenerator {
       case AppRoutes.login:
         return MaterialPageRoute(builder: (_) => LoginPage());
       case AppRoutes.home:
-        return MaterialPageRoute(builder: (_) => HomePage());
+        return MaterialPageRoute(builder: (_) => QuotePage());
       default:
         return MaterialPageRoute(builder: (_) => LoginPage());
     }
