--- lib/features/home/domain/repositories/quote_repository.dart
@@ -1,5 +1,0 @@
-import 'package:challenge_app/features/home/data/models/quote_response.dart';
-
-abstract class QuoteRepository {
-  Future<QuoteResponse> fetchQuote();
-}
