+++ lib/features/quote/data/models/quote_response.dart
@@ -2,7 +2,6 @@
 //
 //     final quoteResponse = quoteResponseFromJson(jsonString);
 
-import 'package:meta/meta.dart';
 import 'dart:convert';
 
 QuoteResponse quoteResponseFromJson(String str) =>
