+++ lib/features/quote/data/services/quote_service_impl.dart
@@ -1,6 +1,6 @@
 import 'package:challenge_app/core/data/api_path.dart';
 import 'package:challenge_app/core/data/http_client.dart';
-import 'package:challenge_app/features/home/domain/services/quote_service.dart';
+import 'package:challenge_app/features/quote/domain/services/quote_service.dart';
 
 class QuoteServiceImpl extends QuoteService {
   HttpClient httpClient;
