+++ lib/features/quote/domain/repositories/quote_repository.dart
@@ -1,0 +1,5 @@
+import 'package:challenge_app/features/quote/data/models/quote_response.dart';
+
+abstract class QuoteRepository {
+  Future<QuoteResponse> fetchQuote();
+}
