+++ lib/features/quote/domain/services/quote_service.dart
