+++ lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:challenge_app/challenge_app.dart';
 import 'package:challenge_app/core/dependency_injector.dart';
 import 'package:challenge_app/core/providers/loading_provider.dart';
-import 'package:challenge_app/features/home/presentation/providers/quote_provider.dart';
+import 'package:challenge_app/features/quote/presentation/providers/quote_provider.dart';
 import 'package:challenge_app/features/login/presentation/providers/biometric_provider.dart';
 import 'package:challenge_app/features/login/presentation/providers/login_provider.dart';
 import 'package:flutter/material.dart';
